Scripts all commented and cooooollll
