+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/CharacterAsset/Scripts/Abilities/General_Abilities/ABILITY_Dash.cs
@@ -25,6 +25,7 @@ namespace EtrasStarterAssets{
 
             if (!enabled || !abilityEnabled)
             {
+                _inputs.dash = false;
                 return;
             }
