+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/PuncherEnemy/Scripts/BaseEnemyScripts/IDamageable.cs
@@ -1,6 +1,5 @@
-﻿//Create an interface with a shared function that can be called
-//By multiple objects
-namespace EtrasStarterAssets{
+﻿namespace EtrasStarterAssets{
+    //This interface is the current mark for if a player can damage an object
     public interface IDamageable<T>
     {
         void TakeDamage(T damage);
