+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/PuncherEnemy/Scripts/FPSTurretDamageHitbox.cs
@@ -4,6 +4,7 @@ namespace EtrasStarterAssets
 {
     public class FPSTurretDamageHitbox : MonoBehaviour, IDamageable<int>
     {
+        //If this hitbox is hit, take damage through idamageable and send the info to the mainscript
         public FPSTurret mainScript;
         public void TakeDamage(int damage)
         {
