+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/PuncherEnemy/Scripts/FPSTurretFist.cs
@@ -1,18 +1,40 @@
-﻿using UnityEngine;
+﻿using System.Collections;
+using UnityEngine;
 
 namespace EtrasStarterAssets
 {
     public class FPSTurretFist : MonoBehaviour
     {
         public FPSTurret mainScript;
-
+        bool cooldown;
         private void OnTriggerEnter(Collider other)
         {
-
+            //If the player is hit, launch them back
             if (other.gameObject.tag == "Player")
             {
                 mainScript.launchPlayer(other.gameObject);
             }
+
+            //If a damageable thing is hit (like another puncher), deal damage to it
+            if (!cooldown)
+            {
+                IDamageable<int> isDamageableCheck = other.gameObject.GetComponent<IDamageable<int>>();
+                if (isDamageableCheck != null)
+                {
+                    isDamageableCheck.TakeDamage(1);
+                    //Wait one second before applying damage again with that fist
+                    StartCoroutine(friendlyFireCooldown());
+                }
+            }
+
         }
+
+
+        IEnumerator friendlyFireCooldown() {
+            cooldown = true;
+            yield return new WaitForSeconds(1);
+            cooldown = false;
+        }
+        
     }
 }
