--- Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/StealthBot.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0ec66cdda392df04aa7f4dc4381103f6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
