+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/TargetRopeAndBridge/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 71de8291c21672e49b33bd6964ad5bda
+guid: 5e09c6da4f76c3a488508cc2c26a99e0
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
