+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/TargetRopeAndBridge/Scripts/RopePoint.cs
@@ -1,0 +1,28 @@
+using UnityEngine;
+
+namespace EtrasStarterAssets
+{
+    public class RopePoint : MonoBehaviour, ISerializationCallbackReceiver
+    {
+        public Target target;
+
+        //This script exists so moving the rope point updates the overall rope line renderer in the editor
+        private void Reset()
+        {
+            target = GetComponentInParent<Target>();
+            OnBeforeSerialize();
+        }
+        public void OnBeforeSerialize()
+        {
+            if (target != null)
+            {
+                target.OnValidate();
+            }
+        }
+
+        public void OnAfterDeserialize()
+        {
+
+        }
+    }
+}
