+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/TargetRopeAndBridge/Scripts/RopePoint.cs.meta
