+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/TargetRopeAndBridge/Scripts/Target.cs.meta
