+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/TargetRopeAndBridge/Scripts/TargetHitbox.cs
@@ -1,0 +1,18 @@
+using UnityEngine;
+
+namespace EtrasStarterAssets
+{
+    public class TargetHitbox : MonoBehaviour, IDamageable<int>
+    {
+        //Recieve damage through the IDamageable interface
+        Target target;
+
+        public void TakeDamage(int damage)
+        {
+            target = GetComponentInParent<Target>();
+            target.targetHit();
+        }
+
+    }
+}
+
