+++ Assets/Etra Games/Etra'sStarterAssets/1-UserAssets/InteractableObjects/WeightedButton/WeightedButton.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
 namespace EtrasStarterAssets{
-    public class Interactable_WeightedButton : MonoBehaviour
+    public class WeightedButton : MonoBehaviour
     {
         //From Just Kris#0001
         /*
