+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Materials/Pickups/PickupOrange.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 0eb3f2f403b4f6d4b8ba059112bf4d82
-folderAsset: yes
-DefaultImporter:
+guid: a57f692cf0e0e0b4680e1698281f7a11
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
