+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Materials/RopeMat.mat
