+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Models/StealthBot.prefab.meta
