+++ Assets/Etra Games/Etra'sStarterAssets/2-Source/Textures/ropeTextureBeta.png.meta
