--- Assets/RopePoint.cs
@@ -1,25 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class RopePoint : MonoBehaviour, ISerializationCallbackReceiver
-{
-     public INTERACTABLE_Target target;
-    private void Reset()
-    {
-        target = GetComponentInParent<INTERACTABLE_Target>();
-        OnBeforeSerialize();
-    }
-    public void OnBeforeSerialize()
-    {
-        if (target != null)
-        {
-           target.OnValidate();
-        }
-    }
-
-    public void OnAfterDeserialize()
-    {
-
-    }
-}
